requirements.txt
Django==3.1.5
djangorestframework==3.12.2
psycopg2-binary==2.8.6
django-cors-headers==3.6.0
PyJWT==2.0.1